pyproject.toml
[tool.poetry]
name = "eclipse24"
version = "0.1.0"
description = "Eclipse24 – Internal Server Decommissioning Utility"
authors = ["<Your-Name> <your-email@company.com>"]
readme = "README.md"
packages = [{ include = "eclipse24", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

fastapi = ">=0.110.0,<0.116.0"
uvicorn = { version = ">=0.29.0,<0.31.0", extras = ["standard"] }
python-dotenv = ">=1.0.0,<2.0.0"
pydantic = ">=2.6.0,<2.9.0"
httpx = ">=0.26.0,<0.28.0"
aiokafka = ">=0.10.0"
aiohttp = ">=3.13.2"
structlog = ">=24.1.0,<25.0.0"
python-multipart = ">=0.0.9"
aio-pika = ">=9.3.0,<10.0.0"
confluent-kafka = ">=2.5.0,<3.0.0"
aws-msk-iam-sasl-signer-python = ">=1.0.2,<2.0.0"
slack_sdk = ">=3.33.0,<4.0.0"
asyncpg = ">=0.29.0,<0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0,<9.0.0"
ruff = ">=0.5.0,<0.6.0"
mypy = ">=1.8.0,<1.11.0"
types-requests = "*"

[[tool.poetry.source]]
name = "prden2-sre-platform-sample-virtual"
url = "<https://your-directory.com>"
priority = "primary"

[tool.poetry.scripts]
# eclipse24-api = "eclipse24.services.api.main:main"
# eclipse24-poller = "eclipse24.services.poller.main:main"
# eclipse24-blocker = "eclipse24.services.blocker.main:main"
# eclipse24-approver = "eclipse24.services.approver.main:main"
# eclipse24-reminder = "eclipse24.services.reminder.main:main"
# eclipse24-validator = "eclipse24.services.validator.main:main"
# eclipse24-expiry-watcher = "eclipse24.services.expiry_watcher.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"